+++ storage/voxel_buffer_gd.h
@@ -133,6 +133,8 @@ public:
 	void set_channel_depth(unsigned int channel_index, Depth new_depth);
 	Depth get_channel_depth(unsigned int channel_index) const;
 
+	void remap_values(unsigned int channel_index, PackedInt32Array map);
+
 	// When using lower than 32-bit resolution for terrain signed distance fields,
 	// it should be scaled to better fit the range of represented values since the storage is normalized to -1..1.
 	// This returns that scale for a given depth configuration.
@@ -160,6 +162,8 @@ public:
 	static Ref<Image> debug_print_sdf_to_image_top_down(const VoxelBufferInternal &vb);
 	Array debug_print_sdf_y_slices(float scale) const;
 	Ref<Image> debug_print_sdf_y_slice(float scale, int y) const;
+	static Ref<Image> debug_print_sdf_y_slice(const VoxelBufferInternal &buffer, float scale, int y);
+	static Ref<Image> debug_print_sdf_z_slice(const VoxelBufferInternal &buffer, float scale, int z);
 
 private:
 	void _b_deprecated_optimize();